refactory: create adapter
